Remove BLUEPRINT.base.file
Fixes a few other imports etc. along the way
