+++ .flake8
@@ -100,9 +100,6 @@ ignore_names =
     *3D*,
     *4D*,
     *5D*,
-    make_BP_path,
-    *get_BP_path*,
-    get_BP_root,
     BB,
     *VV*,
     *TS*,
