+++ BLUEPRINT/base/__init__.py
@@ -1,5 +1,5 @@
 # flake8: noqa
 from bluemira.base.look_and_feel import bluemira_print, print_banner, bluemira_warn
-from .file import make_BP_path, FileManager, SUB_DIRS
+from bluemira.base.file import make_bluemira_path, FileManager, SUB_DIRS
 from bluemira.base.file import get_files_by_ext
 from .palettes import BLUE
