+++ documentation/BLUEPRINT/source/base.rst
@@ -74,7 +74,7 @@ Absolute & Relative Paths
 * `generated_data_root` is used to store output data generated by the run.
 
 These paths may be absolute, or they may be relative to the BLUEPRINT root folder by prefixing
-the keyword ``!BP_ROOT!``. When the `Reactor` class is instantiated, the keyword will
+the keyword ``!BM_ROOT!``. When the `Reactor` class is instantiated, the keyword will
 be replaced with the path to BLUEPRINT root for the current system.
 
 For example:
@@ -82,8 +82,8 @@ For example:
 .. code-block:: pycon
 
     >>> Build_Config = {
-            "reference_data_root": "!BP_ROOT!/data",
-            "generated_data_root": "!BP_ROOT!/generated_data",
+            "reference_data_root": "!BM_ROOT!/data",
+            "generated_data_root": "!BM_ROOT!/generated_data",
         }
 
 Type-checking
