+++ examples/BLUEPRINT/ST.py
@@ -23,15 +23,14 @@ A typical spherical tokamak fusion power reactor.
 """
 import os
 
-from BLUEPRINT.base.file import make_BP_path
-from bluemira.base.file import get_bluemira_root
+from bluemira.base.file import BM_ROOT, get_bluemira_root, make_bluemira_path
 from bluemira.base.look_and_feel import plot_defaults, print_banner
 from BLUEPRINT.reactor import Reactor
 from BLUEPRINT.systems.config import Spherical
 
 plot_defaults()
 KEY_TO_PLOT = False
-PLOTFOLDER = make_BP_path("Data/plots")
+PLOTFOLDER = make_bluemira_path("Data/plots")
 if os.path.isdir(PLOTFOLDER) is False:
     KEY_TO_PLOT = False
 
@@ -64,7 +63,7 @@ config = {
 }
 
 build_config = {
-    "generated_data_root": "!BP_ROOT!/generated_data/BLUEPRINT",
+    "generated_data_root": f"{BM_ROOT}/generated_data/BLUEPRINT",
     "plot_flag": False,
     "process_mode": "mock",
     "process_indat": os.path.join(
