+++ examples/BLUEPRINT/cad/cad_tutorial.py
@@ -34,7 +34,7 @@ Feel free to change parameters!
 
 # %%
 import os
-from BLUEPRINT.base.file import get_BP_path
+from bluemira.base.file import get_bluemira_path
 from BLUEPRINT.geometry.loop import Loop
 from BLUEPRINT.cad.cadtools import (
     make_face,
@@ -205,7 +205,7 @@ show_CAD(cutlet)
 # For this we're going to load some Loop shapes from files
 
 # %%
-path = get_BP_path("BLUEPRINT/cad/test_data", subfolder="tests")
+path = get_bluemira_path("BLUEPRINT/cad/test_data", subfolder="tests")
 name = "plasmaloop.json"
 filename = os.sep.join([path, name])
 
@@ -271,7 +271,7 @@ filename = os.sep.join([path, "plasma_block_test"])
 save_as_STEP(plasma_block2, filename)
 
 # %%[markdown]
-# You can check it was saved by typing: `exporer.exe .` in your ubuntu terminal, and
+# You can check it was saved by typing: `explorer.exe .` in your ubuntu terminal, and
 # navigating to the file path above.
 
 # You can view the file in any CAD program, like FreeCAD..
