+++ examples/BLUEPRINT/nova/superconducting_tf_coil_example.py
@@ -29,7 +29,7 @@ import matplotlib.pyplot as plt
 
 from bluemira.base.parameter import ParameterFrame
 
-from BLUEPRINT.base.file import make_BP_path
+from bluemira.base.file import make_bluemira_path
 from BLUEPRINT.geometry.loop import Loop
 from BLUEPRINT.systems.tfcoils import ToroidalFieldCoils
 from bluemira.equilibria.shapes import flux_surface_manickam
@@ -59,8 +59,8 @@ params = [
 
 parameters = ParameterFrame(params)
 
-read_path = make_BP_path("Geometry", subfolder="data/BLUEPRINT")
-write_path = make_BP_path("SC_P_coil", subfolder="generated_data/BLUEPRINT")
+read_path = make_bluemira_path("Geometry", subfolder="data/BLUEPRINT")
+write_path = make_bluemira_path("SC_P_coil", subfolder="generated_data/BLUEPRINT")
 
 lcfs = flux_surface_manickam(3.639, 0, 2.183, 2.8, 0.543, n=40)
 lcfs.close()
