+++ tests/BLUEPRINT/systems/test_plotting.py
@@ -25,7 +25,7 @@ import pytest
 from matplotlib import pyplot as plt
 import filecmp
 
-from BLUEPRINT.base.file import FileManager
+from bluemira.base.file import FileManager
 from BLUEPRINT.nova.structure import CoilArchitect
 from BLUEPRINT.systems.blanket import BreedingBlanket
 from BLUEPRINT.systems.buildings import RadiationShield
