+++ tests/bluemira/base/test_file.py
@@ -20,7 +20,19 @@
 # License along with bluemira; if not, see <https://www.gnu.org/licenses/>.
 
 import pytest
-from bluemira.base.file import get_bluemira_path, try_get_bluemira_path
+
+import os
+
+from bluemira.base.file import (
+    FileManager,
+    get_bluemira_path,
+    SUB_DIRS,
+    try_get_bluemira_path,
+)
+
+REACTOR_NAME = "TEST_REACTOR"
+FILE_REF_PATH = os.path.join("data")
+FILE_GEN_PATH = os.path.join("tests", "bluemira", "base", "file_manager_gen")
 
 
 @pytest.mark.parametrize(
@@ -52,6 +64,182 @@ def test_try_get_bluemira_path(path, subfolder, allow_missing, expect_none):
 def test_try_get_bluemira_path_raises(path, subfolder):
     with pytest.raises(ValueError):
         try_get_bluemira_path(path, subfolder=subfolder, allow_missing=False)
+
+
+@pytest.fixture(scope="session", autouse=True)
+def file_manager_good():
+    """
+    Create a FileManager to run some tests on.
+    """
+    file_manager = FileManager(
+        REACTOR_NAME,
+        reference_data_root=FILE_REF_PATH,
+        generated_data_root=FILE_GEN_PATH,
+    )
+
+    yield file_manager
+
+    # Make sure we clean up the directories after testing that they have been created.
+    if os.path.exists(FILE_GEN_PATH):
+        remove_dir_and_subs(FILE_GEN_PATH)
+
+
+@pytest.fixture(scope="session", autouse=True)
+def file_manager_bad():
+    """
+    Create a FileManager to run some tests on.
+    """
+    file_manager = FileManager(
+        REACTOR_NAME,
+        reference_data_root=FILE_REF_PATH + "bad",
+        generated_data_root=FILE_GEN_PATH,
+    )
+
+    return file_manager
+
+
+def assert_bluemira_path_exists(path: str):
+    """
+    Fails a test if the specified path does not exist.
+    """
+    try:
+        get_bluemira_path(path, "")
+    except ValueError as error:
+        pytest.fail("{path} does not exist and has not been created: %s" % error)
+
+
+def remove_dir_and_subs(path: str):
+    """
+    Removes the specified path and any empty sub-directories.
+    """
+    for root, dirs, _ in os.walk(get_bluemira_path(path, ""), topdown=False):
+        for name in dirs:
+            os.rmdir(os.path.join(root, name))
+    os.rmdir(get_bluemira_path(path, ""))
+
+
+def get_system_path(root: str, reactor_name: str, system: str) -> str:
+    """
+    Builds the path for a system given a root directory and a reactor
+    """
+    return os.path.join(root, "reactors", reactor_name, system)
+
+
+def get_reactor_path(root: str, reactor_name: str) -> str:
+    """
+    Builds the path for a reactor given a root directory and a reactor
+    """
+    return os.path.join(root, "reactors", reactor_name)
+
+
+def test_create_generated_data_root(file_manager_good: FileManager):
+    """
+    Tests that the generated data root directories are created.
+    """
+    file_manager_good.build_dirs()
+
+    assert_bluemira_path_exists(file_manager_good.generated_data_root)
+
+
+def test_set_reference_data_root(file_manager_good: FileManager):
+    """
+    Tests that the reference data root directories are set.
+    """
+    file_manager_good.build_dirs()
+
+    assert_bluemira_path_exists(file_manager_good.reference_data_root)
+
+
+def test_reference_data_dirs(file_manager_good: FileManager):
+    """
+    Tests that the reference data path dictionary is created and populated.
+    """
+    file_manager_good.set_reference_data_paths()
+
+    target_data_dirs = sorted(SUB_DIRS + ["root"])
+
+    # Check the keys match the expected list
+    assert sorted(file_manager_good.reference_data_dirs.keys()) == target_data_dirs
+
+    for key, value in file_manager_good.reference_data_dirs.items():
+        if key == "root":
+            assert value == get_reactor_path(
+                file_manager_good.reference_data_root, file_manager_good.reactor_name
+            )
+        else:
+            assert value == get_system_path(
+                file_manager_good.reference_data_root,
+                file_manager_good.reactor_name,
+                key,
+            )
+
+
+def test_reference_data_dirs_error(file_manager_bad: FileManager):
+    """
+    Tests that a ValueError is raised if the data path does not exist.
+    """
+    with pytest.raises(ValueError):
+        file_manager_bad.set_reference_data_paths()
+
+
+def test_create_reference_data_dirs():
+    """
+    Test that the reference data path dictionary is created and populated
+
+    Creates the directory and removes it.
+    """
+    file_manager = FileManager(
+        REACTOR_NAME,
+        reference_data_root=FILE_REF_PATH + "create",
+        generated_data_root=FILE_GEN_PATH,
+    )
+
+    file_manager.create_reference_data_paths()
+
+    target_data_dirs = sorted(SUB_DIRS + ["root"])
+
+    # Check the keys match the expected list
+    assert sorted(file_manager.reference_data_dirs.keys()) == target_data_dirs
+
+    for key, value in file_manager.reference_data_dirs.items():
+        if key == "root":
+            assert value == get_reactor_path(
+                file_manager.reference_data_root, file_manager.reactor_name
+            )
+        else:
+            assert value == get_system_path(
+                file_manager.reference_data_root,
+                file_manager.reactor_name,
+                key,
+            )
+
+    # Make sure we clean up the directories after testing that they have been created.
+    if os.path.exists(FILE_GEN_PATH):
+        remove_dir_and_subs(FILE_GEN_PATH)
+
+
+def test_generated_data_dirs(file_manager_good: FileManager):
+    """
+    Tests that the generated data path dictionary is created and populated.
+    """
+    file_manager_good.create_generated_data_paths()
+
+    target_data_dirs = sorted(SUB_DIRS + ["root"])
+
+    # Check the keys match the expected list
+    assert sorted(file_manager_good.generated_data_dirs.keys()) == target_data_dirs
+
+    for key, value in file_manager_good.generated_data_dirs.items():
+        if key == "root":
+            assert value == get_reactor_path(
+                file_manager_good.generated_data_root, file_manager_good.reactor_name
+            )
+        else:
+            assert value == get_system_path(
+                file_manager_good.generated_data_root,
+                file_manager_good.reactor_name,
+                key,
+            )
 
 
 if __name__ == "__main__":
